requirements.txt
selenium==4.20.0
python-dotenv